requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Cython==0.29.23
distro==1.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
docopt==0.6.2
drf-yasg==1.20.0
fire==0.4.0
fonttools==4.24.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.20.3
openapi-codec==1.3.2
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
pdf2docx==0.5.2
Pillow==8.2.0
pipreqs==0.4.10
pycparser==2.20
PyJWT==2.1.0
PyMuPDF==1.18.14
PyMySQL==1.0.2
pypandoc==1.5
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.7
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tabula-py==2.2.0
termcolor==1.1.0
tika==1.24
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
yarg==0.1.9
